switch-display: hack fix for monitor WS assignment
Keep having hot plug issues with my main monitor, VRR and slow wakeup
results in workspaces being shifted to faster wakeup monitors. I need
most ws to be on main, so this hack will shift all workspace back. It's
not perfect but it's better than not doing it. Hopefully there will be
an upstream fix for this.