switch-display: fix WS monitor assigment
Fixes issue where workspaces was moved to a different monitor than what
assigned in config. The reason is that my displays all have a slightly
different wakeup time, and hyprland will rearrange ws on display
changes and if a display isn't available (yet) it will move it to the
best available. Calling reload again will not fix this, so I had to
force it.

This disabled toggle, I don't use it, re-implement if needed.